requirements.txt
ultralytics
opencv-python
pillow
scikit-learn
pyyaml
torch
torchvision
